PDFBOX-2576: remove unneeded casts
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1666325 13f79535-47bb-0310-9956-ffa450edef68

+++ preflight/src/main/java/org/apache/pdfbox/preflight/metadata/SynchronizedMetaDataValidation.java
@@ -582,7 +582,7 @@ public class SynchronizedMetaDataValidation
     {
         // remove trailing NUL values
         int length = string.length();
-        while(length > 0 && (int)string.charAt(length-1) == 0)
+        while (length > 0 && string.charAt(length - 1) == 0)
         {
             length--;
         }
